Implement per parameter CommandCompletion
A `@CommandCompletion` annotation can be provided per parameter to provide tab completions specific to that parameter. This is useful if a parameter consumes multiple arguments and is required for named parameters to be implemented. It also provides the framework for a context manager to provide its own default completions.

Backwards compatiblity is maintained by merging in CommandCompletions provided on the method.

Changes:
  * Update CommandCompletion annotation to work on Parameters
  * Provide helper methods in CommandParameter to access the completions
  * Provide a helper method in CommandCompletions to merge a method command completion and the parameter completions.